Organize frontend file structure
