+++ app/frontend/src/components/Buttons/InviteToTeamButton.tsx
@@ -12,25 +12,21 @@ import {
   ListItem,
   Box
 } from "@mui/material";
-import PersonAddIcon from "@mui/icons-material/PersonAdd";
+import GroupAddIcon from "@mui/icons-material/GroupAdd";
 
 import { useState, useRef, useEffect } from "react";
 
 import wretch from "wretch";
 import { toast } from "react-toastify";
-import { useApplicationStore } from "../stores/ApplicationStore";
-import { API_URL } from "../utils/FetchUtils";
+import { useApplicationStore } from "../../stores/ApplicationStore";
+import { API_URL } from "../../utils/FetchUtils";
 
-interface IInviteToChannelButtonProps {
+interface IInviteToTeamButtonProps {
   teamId: number;
-  channelId: number;
-  channelName: string;
-  refetchData: () => void;
+  teamName: string;
 }
 
-export default function InviteToChannelButton(
-  props: IInviteToChannelButtonProps,
-) {
+export default function InviteToTeamButton(props: IInviteToTeamButtonProps) {
   const [isDialogOpen, setIsDialogOpen] = useState<boolean>(false);
   const [inviteeNames, setInviteeNames] = useState<string[]>([]);
   const [currentUserName, setCurrentUserName] = useState<string>("");
@@ -39,7 +35,8 @@ export default function InviteToChannelButton(
   const [position, setPosition] = useState({top: 0, left: 0, width: 0});
   const searchBarRef = useRef<HTMLInputElement>(null);
   const [alreadyOpen, setAlreadyOpen] = useState<boolean>(false);
-
+  const refetchData = useApplicationStore((state) => state.refetchTeamChannelsState);
+  
   useEffect(() => {
     if (isDialogOpen && !alreadyOpen) {
       getUsers();
@@ -57,27 +54,24 @@ export default function InviteToChannelButton(
   }, [currentUserName, isDialogOpen, showSuggestions]);
 
   const getUsers = () => {
-      wretch(`${API_URL}/api/add/sendteamusers`)
-        .auth(`Bearer ${localStorage.getItem("jwt-token")}`)
-        .post([props.teamId, props.channelId])
-        .json((sug) => {setSuggestions(sug);})
-        .catch((error) => {
-          console.error(error);
-          toast.error("An error has occurred.");
+    wretch(`${API_URL}/api/add/sendusers`)
+      .auth(`Bearer ${localStorage.getItem("jwt-token")}`)
+      .headers({ "Content-Type": "application/json" })
+      .post(JSON.stringify(props.teamId))
+      .json((sug) => {setSuggestions(sug);})
+      .catch((error) => {
+        console.error(error);
+        toast.error("An error has occurred.");
       });
   };
 
   const onSubmit = () => {
     if (inviteeNames.length > 0) {
-      wretch(`${API_URL}/api/add/addtochannel`)
+      wretch(`${API_URL}/api/add/addtoteam`)
         .auth(`Bearer ${localStorage.getItem("jwt-token")}`)
-        .post({
-          team_id: props.teamId,
-          channel_id: props.channelId,
-          users_to_add: inviteeNames,
-        })
+        .post({ team_id: props.teamId, users_to_add: inviteeNames })
         .res(() => {
-          props.refetchData();
+          refetchData();
           toast.success("User" + (inviteeNames.length > 1 ? "s have" : " has") + " been added successfully.");
           quit();
         })
@@ -98,8 +92,8 @@ export default function InviteToChannelButton(
 
   return (
     <>
-      <Dialog open={isDialogOpen} onClose={quit}>
-        <DialogTitle>Add Users to {props.channelName}</DialogTitle>
+      <Dialog open={isDialogOpen} onClose={() => quit()}>
+        <DialogTitle>Add Users to {props.teamName}</DialogTitle>
         <DialogContent
           sx={{
             minHeight: "100px",
@@ -181,13 +175,12 @@ export default function InviteToChannelButton(
           </Button>
         </DialogActions>
       </Dialog>{" "}
-      <Tooltip title="Assign users to this channel">
+      <Tooltip title="Assign users to the team">
         <IconButton
-          sx={{ maxHeight: "24px", borderRadius: "4px" }}
-          edge="end"
+          sx={{ height: "52px", width: "47%" }}
           onClick={() => setIsDialogOpen(true)}
         >
-          <PersonAddIcon />
+          <GroupAddIcon></GroupAddIcon>
         </IconButton>
       </Tooltip>
     </>
